pipelines/ai/requirements.txt
litellm>=1.40
pyyaml>=6.0
pandas>=2.2
